+++ crates/mdbook-markdown/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "mdbook-markdown"
+version = "0.5.0-alpha.1"
+description = "Markdown processing used in mdBook"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+log.workspace = true
+pulldown-cmark.workspace = true
+regex.workspace = true
+
+[lints]
+workspace = true
