+++ src/renderer/html_handlebars/hbs_renderer.rs
@@ -17,6 +17,8 @@ use log::{debug, info, trace, warn};
 use mdbook_core::config::{BookConfig, Code, Config, HtmlConfig, Playground, RustEdition};
 use mdbook_core::utils;
 use mdbook_core::utils::fs::get_404_output_file;
+use mdbook_markdown::{render_markdown, render_markdown_with_path};
+
 use regex::{Captures, Regex};
 use serde_json::json;
 
@@ -57,13 +59,10 @@ impl HtmlHandlebars {
                 .insert("git_repository_edit_url".to_owned(), json!(edit_url));
         }
 
-        let content = utils::render_markdown(&ch.content, ctx.html_config.smart_punctuation());
+        let content = render_markdown(&ch.content, ctx.html_config.smart_punctuation());
 
-        let fixed_content = utils::render_markdown_with_path(
-            &ch.content,
-            ctx.html_config.smart_punctuation(),
-            Some(path),
-        );
+        let fixed_content =
+            render_markdown_with_path(&ch.content, ctx.html_config.smart_punctuation(), Some(path));
         if !ctx.is_index && ctx.html_config.print.page_break {
             // Add page break between chapters
             // See https://developer.mozilla.org/en-US/docs/Web/CSS/break-before and https://developer.mozilla.org/en-US/docs/Web/CSS/page-break-before
@@ -178,8 +177,7 @@ impl HtmlHandlebars {
                     .to_string()
             }
         };
-        let html_content_404 =
-            utils::render_markdown(&content_404, html_config.smart_punctuation());
+        let html_content_404 = render_markdown(&content_404, html_config.smart_punctuation());
 
         let mut data_404 = data.clone();
         let base_url = if let Some(site_url) = &html_config.site_url {
