+++ src/renderer/html_handlebars/search.rs
@@ -8,6 +8,7 @@ use elasticlunr::{Index, IndexBuilder};
 use log::{debug, warn};
 use mdbook_core::config::{Search, SearchChapterSettings};
 use mdbook_core::utils;
+use mdbook_markdown::new_cmark_parser;
 use pulldown_cmark::*;
 use serde::Serialize;
 
@@ -134,7 +135,7 @@ fn render_item(
         .with_context(|| "Could not convert HTML path to str")?;
     let anchor_base = utils::fs::normalize_path(filepath);
 
-    let mut p = utils::new_cmark_parser(&chapter.content, false).peekable();
+    let mut p = new_cmark_parser(&chapter.content, false).peekable();
 
     let mut in_heading = false;
     let max_section_depth = u32::from(search_config.heading_split_level);
